[rename]: 스탬프북 슬라이더 이름 변경
