+++ src/features/main/ui/StampSlider.tsx
